feat: add remove use-get-user, add get-user
